Isolate MergeTree to stop export
