+++ api-report/merge-tree.api.md
@@ -25,6 +25,8 @@ export interface AugmentedIntervalNode {
 
 // @public (undocumented)
 export abstract class BaseSegment extends MergeNode implements ISegment {
+    // Warning: (ae-forgotten-export) The symbol "MergeTree" needs to be exported by the entry point index.d.ts
+    //
     // (undocumented)
     ack(segmentGroup: SegmentGroup, opArgs: IMergeTreeDeltaOpArgs, mergeTree: MergeTree): boolean;
     // (undocumented)
@@ -185,7 +187,7 @@ export class Client {
     maxWindowTime: number;
     // (undocumented)
     measureOps: boolean;
-    // (undocumented)
+    // @internal @deprecated (undocumented)
     protected readonly mergeTree: MergeTree;
     // (undocumented)
     get mergeTreeDeltaCallback(): MergeTreeDeltaCallback | undefined;
@@ -219,21 +221,6 @@ export class Client {
     // (undocumented)
     walkSegments<undefined>(handler: ISegmentAction<undefined>, start?: number, end?: number, accum?: undefined, splitRange?: boolean): void;
 }
-
-// Warning: (ae-internal-missing-underscore) The name "ClientSeq" should be prefixed with an underscore because the declaration is marked as @internal
-//
-// @internal @deprecated (undocumented)
-export interface ClientSeq {
-    // (undocumented)
-    clientId: string;
-    // (undocumented)
-    refSeq: number;
-}
-
-// Warning: (ae-internal-missing-underscore) The name "clientSeqComparer" should be prefixed with an underscore because the declaration is marked as @internal
-//
-// @internal @deprecated (undocumented)
-export const clientSeqComparer: Comparer<ClientSeq>;
 
 // @public (undocumented)
 export function clone<T>(extension: MapLike<T> | undefined): MapLike<T> | undefined;
@@ -890,16 +877,6 @@ export interface LocalReferencePosition extends ReferencePosition {
     callbacks?: Partial<Record<"beforeSlide" | "afterSlide", () => void>>;
 }
 
-// Warning: (ae-internal-missing-underscore) The name "LRUSegment" should be prefixed with an underscore because the declaration is marked as @internal
-//
-// @internal @deprecated (undocumented)
-export interface LRUSegment {
-    // (undocumented)
-    maxSeq: number;
-    // (undocumented)
-    segment?: ISegment;
-}
-
 // @public (undocumented)
 export interface MapLike<T> {
     // (undocumented)
@@ -963,10 +940,8 @@ export class MergeBlock extends MergeNode implements IMergeBlock {
     childCount: number;
     // (undocumented)
     children: IMergeNode[];
-    // Warning: (ae-forgotten-export) The symbol "HierMergeBlock" needs to be exported by the entry point index.d.ts
-    //
     // (undocumented)
-    hierBlock(): HierMergeBlock | undefined;
+    hierBlock(): IHierBlock | undefined;
     // (undocumented)
     setOrdinal(child: IMergeNode, index: number): void;
 }
@@ -983,100 +958,6 @@ export class MergeNode implements IMergeNodeCommon {
     ordinal: string;
     // (undocumented)
     parent?: IMergeBlock;
-}
-
-// @public (undocumented)
-export class MergeTree {
-    constructor(options?: PropertySet | undefined);
-    ackPendingSegment(opArgs: IMergeTreeDeltaOpArgs): void;
-    // (undocumented)
-    addMinSeqListener(minRequired: number, onMinGE: (minSeq: number) => void): void;
-    annotateRange(start: number, end: number, props: PropertySet, combiningOp: ICombiningOp | undefined, refSeq: number, clientId: number, seq: number, opArgs: IMergeTreeDeltaOpArgs): void;
-    // (undocumented)
-    blockClone(block: IMergeBlock, segments?: ISegment[]): MergeBlock;
-    // (undocumented)
-    clone(): void;
-    // (undocumented)
-    readonly collabWindow: CollaborationWindow;
-    // (undocumented)
-    createLocalReferencePosition(segment: ISegment, offset: number, refType: ReferenceType, properties: PropertySet | undefined): LocalReferencePosition;
-    // (undocumented)
-    findTile(startPos: number, clientId: number, tileLabel: string, posPrecedesTile?: boolean): {
-        tile: ReferencePosition;
-        pos: number;
-    } | undefined;
-    // (undocumented)
-    getCollabWindow(): CollaborationWindow;
-    // (undocumented)
-    getContainingSegment<T extends ISegment>(pos: number, refSeq: number, clientId: number): {
-        segment: T | undefined;
-        offset: number | undefined;
-    };
-    // (undocumented)
-    getLength(refSeq: number, clientId: number): number;
-    // (undocumented)
-    getMarkerFromId(id: string): ISegment | undefined;
-    // (undocumented)
-    getPosition(node: MergeNode, refSeq: number, clientId: number): number;
-    // @internal
-    _getSlideToSegment(segoff: {
-        segment: ISegment | undefined;
-        offset: number | undefined;
-    }): {
-        segment: ISegment | undefined;
-        offset: number | undefined;
-    };
-    // @internal @deprecated (undocumented)
-    getStackContext(startPos: number, clientId: number, rangeLabels: string[]): RangeStackMap;
-    // (undocumented)
-    getStats(): MergeTreeStats;
-    // @internal @deprecated (undocumented)
-    incrementalBlockMap<TContext>(stateStack: Stack<IncrementalMapState<TContext>>): void;
-    // (undocumented)
-    insertAtReferencePosition(referencePosition: ReferencePosition, insertSegment: ISegment, opArgs: IMergeTreeDeltaOpArgs): void;
-    // (undocumented)
-    insertSegments(pos: number, segments: ISegment[], refSeq: number, clientId: number, seq: number, opArgs: IMergeTreeDeltaOpArgs | undefined): void;
-    get length(): number;
-    // (undocumented)
-    localNetLength(segment: ISegment): number;
-    // (undocumented)
-    map<TClientData>(actions: SegmentActions<TClientData>, refSeq: number, clientId: number, accum: TClientData): void;
-    // (undocumented)
-    mapIdToSegment(id: string, segment: ISegment): void;
-    // (undocumented)
-    mapRange<TClientData>(actions: SegmentActions<TClientData>, refSeq: number, clientId: number, accum: TClientData, start?: number, end?: number, splitRange?: boolean): void;
-    // (undocumented)
-    markRangeRemoved(start: number, end: number, refSeq: number, clientId: number, seq: number, overwrite: boolean | undefined, opArgs: IMergeTreeDeltaOpArgs): void;
-    // (undocumented)
-    mergeTreeDeltaCallback?: MergeTreeDeltaCallback;
-    // (undocumented)
-    mergeTreeMaintenanceCallback?: MergeTreeMaintenanceCallback;
-    // (undocumented)
-    options?: PropertySet | undefined;
-    // (undocumented)
-    static readonly options: {
-        incrementalUpdate: boolean;
-        insertAfterRemovedSegs: boolean;
-        zamboniSegments: boolean;
-    };
-    // @internal @deprecated (undocumented)
-    pendingSegments: List<SegmentGroup> | undefined;
-    posFromRelativePos(relativePos: IRelativePosition, refseq?: number, clientId?: number): number;
-    // (undocumented)
-    referencePositionToLocalPosition(refPos: ReferencePosition, refSeq?: number, clientId?: number): number;
-    // (undocumented)
-    reloadFromSegments(segments: ISegment[]): void;
-    // (undocumented)
-    removeLocalReferencePosition(lref: LocalReferencePosition): LocalReferencePosition | undefined;
-    resolveRemoteClientPosition(remoteClientPosition: number, remoteClientRefSeq: number, remoteClientId: number): number | undefined;
-    // (undocumented)
-    root: IMergeBlock;
-    // (undocumented)
-    setMinSeq(minSeq: number): void;
-    // (undocumented)
-    startCollaboration(localClientId: number, minSeq: number, currentSeq: number): void;
-    // (undocumented)
-    walkAllSegments<TClientData>(block: IMergeBlock, action: (segment: ISegment, accum?: TClientData) => boolean, accum?: TClientData): boolean;
 }
 
 // @public (undocumented)
