+++ packages/dds/merge-tree/src/test/mergeTree.annotate.deltaCallback.spec.ts
@@ -4,11 +4,11 @@
  */
 
 import { strict as assert } from "assert";
-import { MergeTree } from "../mergeTree";
 import { MergeTreeDeltaType } from "../ops";
 import { MergeTreeMaintenanceType } from "../mergeTreeDeltaCallback";
 import { LocalClientId, UnassignedSequenceNumber, UniversalSequenceNumber } from "../constants";
 import { TextSegment } from "../textSegment";
+import { MergeTree } from "../mergeTreeInt";
 import { countOperations, insertText } from "./testUtils";
 
 describe("MergeTree", () => {
