+++ packages/dds/merge-tree/src/test/testClient.ts
@@ -15,12 +15,13 @@ import {
     ListMakeHead,
 } from "../collections";
 import { UnassignedSequenceNumber } from "../constants";
-import { ISegment, Marker, MergeTree } from "../mergeTree";
+import { ISegment, Marker } from "../mergeTree";
 import { createInsertSegmentOp, createRemoveRangeOp } from "../opBuilder";
 import { IJSONSegment, IMarkerDef, IMergeTreeOp, MergeTreeDeltaType, ReferenceType } from "../ops";
 import { PropertySet } from "../properties";
 import { SnapshotLegacy } from "../snapshotlegacy";
 import { MergeTreeTextHelper, TextSegment } from "../textSegment";
+import { MergeTree } from "../mergeTreeInt";
 import { TestSerializer } from "./testSerializer";
 import { nodeOrdinalsHaveIntegrity } from "./testUtils";
