+++ packages/dds/merge-tree/src/test/testUtils.ts
@@ -8,12 +8,12 @@ import fs from "fs";
 import {
     IMergeBlock,
     Marker,
-    MergeTree,
 } from "../mergeTree";
 import { IMergeTreeDeltaOpArgs } from "../mergeTreeDeltaCallback";
 import { TextSegment } from "../textSegment";
 import { ReferenceType } from "../ops";
 import { PropertySet } from "../properties";
+import { MergeTree } from "../mergeTreeInt";
 import { loadText } from "./text";
 
 export function loadTextFromFile(filename: string, mergeTree: MergeTree, segLimit = 0) {
